Use DeepSeq to better communicate our intentions
Our test suite was full of code like

    do x <- foo
       return (length (show x) > 0)

The reason for this is that `foo >> return True` might be too lazy: `foo` may
return a lazy value which is full of errors, but we'll never trigger those
errors because we're not forcing the result.

There are basically three problems:

  1. `x` may be a compound datastructure, so we need a way to force *all* of it,
    since WHNF may not be enough.
  2. We need the test result to depend on the forced value
  3. We need an approach that is as generic as possible (i.e. cares as little
    about what `x` is as possible)

We solve point 3 by calling `show x`: as long as `x` has a Show instance, this
lets us ignore the type of `x` and just handle a String.

We solve point 1 by calling `length`, which returns a single Int which, when
forced to WHNF, will trigger the whole (spine of the) string to be evaluated,
which (assuming a reasonable `show` implementation) is enough to force
everything in `x`.

We solve point 2 by having our test result be `length ... > 0`, whose value
requires us to fully evaluate `x`.

Note that this is actually a pretty crappy situation: we're using implementation
details to hack our way to a fragile solution. Imagine that some bright spark
implements an optimisation, such that `length` and `>` get fused: for example,
`length` may be implemented by calling `lengthNat` and converting the result to
an Int; we may then spot that

    natToInt (lengthNat ...) `(> :: Int -> Int -> Bool)` 0

can be optimised to:

    lengthNat ... `(> :: Nat -> Nat -> Bool)` 0

The call to `> :: Nat -> Nat -> Bool` can short-circuit as soon as we evaluate
the outer constructor of `lengthNat ...`, since `Z > Z` is False and `S ... > Z`
is always True.

Hence we're fighting the compiler, because we didn't actually tell it what we
wanted! Also note that HLint actually points this out to us: it tells us that
`length ... > 0` should be replaced by `not (null ...)`; the reason we weren't
using `not (null ...)` is that `null` is lazy in exactly the same way that the
above hypothetical optimisation would be lazy!

What did we do with this knowledge? We added annotations to prevent HLint
complaining about these uses. That's a pretty bad smell. Instead we should write
code which communicates, to human readers and automated tooling, what we
*actually* want: to fully evaluate `x` and, if that doesn't trigger an error, to
return True.

We can do this using `deepseq`. For all we know, `deepseq` may be implemented by
comparing the length of a list; it doesn't matter: *semantically*, the meaning
is conveyed by the fact we're using `deepseq`. If such hypothetical
optimisations exist in the future, such that `deepseq` no longer forces the
given value, then that would be a legitimate bug. Interestingly, `deepseq` is an
important-enough function that breaking it would probably be of concern to those
implementing such an optimisation. The result would probably be a new version of
`deepseq` being released alongside the compiler optimisation's rollout, to
ensure that everyone's code keeps working *as intended* (rather than *as
implemented*).

Thankfully, by using deepseq, we don't need to care about such things, since
neither of them (the potential optimisation and the implementation of deepseq)
are our responsibility!